// src/zero/schema.ts
// @generated drizzle-zero signature:sha256:0ce15137f1d4c76cd04daf7d22991e8170f4c26b82933692a2aae614eaa73a0b
// This file was automatically generated by drizzle-zero.
// You should NOT make any changes in this file as it will be overwritten.

import type { Row } from "@rocicorp/zero";
import { createBuilder } from "@rocicorp/zero";

const tasksTable = {
  name: "tasks",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    title: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    completedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    scheduledAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    retryNumber: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    createdAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    updatedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
  },
  primaryKey: ["id"],
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
  tables: {
    tasks: tasksTable,
  },
  relationships: {},
  enableLegacyQueries: false,
  enableLegacyMutators: false,
} as const;

/**
 * Represents the Zero schema type.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Schema = typeof schema;
/**
 * Represents a row from the "tasks" table.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Task = Row<(typeof schema)["tables"]["tasks"]>;

/**
 * Represents the ZQL query builder.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const zql = createBuilder(schema);
/**
 * Represents the Zero schema query builder.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const builder = zql;

/** Defines the default types for Zero */
declare module "@rocicorp/zero" {
  interface DefaultTypes {
    schema: Schema;
  }
}